moves lahman file back to top level
