+++ pybaseball/__init__.py
@@ -24,35 +24,35 @@ from .team_pitching import team_pitching
 from .team_pitching import team_pitching_bref
 from .top_prospects import top_prospects
 from .amateur_draft import amateur_draft
-from .datasources.lahman import parks
-from .datasources.lahman import all_star_full
-from .datasources.lahman import appearances
-from .datasources.lahman import awards_managers
-from .datasources.lahman import awards_players
-from .datasources.lahman import awards_share_managers
-from .datasources.lahman import awards_share_players
-from .datasources.lahman import batting
-from .datasources.lahman import batting_post
-from .datasources.lahman import college_playing
-from .datasources.lahman import fielding
-from .datasources.lahman import fielding_of
-from .datasources.lahman import fielding_of_split
-from .datasources.lahman import fielding_post
-from .datasources.lahman import hall_of_fame
-from .datasources.lahman import home_games
-from .datasources.lahman import managers
-from .datasources.lahman import managers_half
-from .datasources.lahman import master
-from .datasources.lahman import people
-from .datasources.lahman import pitching
-from .datasources.lahman import pitching_post
-from .datasources.lahman import salaries
-from .datasources.lahman import schools
-from .datasources.lahman import series_post
-from .datasources.lahman import teams
-from .datasources.lahman import teams_franchises
-from .datasources.lahman import teams_half
-from .datasources.lahman import download_lahman
+from .lahman import parks
+from .lahman import all_star_full
+from .lahman import appearances
+from .lahman import awards_managers
+from .lahman import awards_players
+from .lahman import awards_share_managers
+from .lahman import awards_share_players
+from .lahman import batting
+from .lahman import batting_post
+from .lahman import college_playing
+from .lahman import fielding
+from .lahman import fielding_of
+from .lahman import fielding_of_split
+from .lahman import fielding_post
+from .lahman import hall_of_fame
+from .lahman import home_games
+from .lahman import managers
+from .lahman import managers_half
+from .lahman import master
+from .lahman import people
+from .lahman import pitching
+from .lahman import pitching_post
+from .lahman import salaries
+from .lahman import schools
+from .lahman import series_post
+from .lahman import teams
+from .lahman import teams_franchises
+from .lahman import teams_half
+from .lahman import download_lahman
 from .retrosheet import season_game_logs
 from .retrosheet import world_series_logs
 from .retrosheet import all_star_game_logs
